Explicitly set return type to Nil
I was seeing a flickering background whenever moving the mouse on part 7.
Adding the nils resolves the issue.
This might be a bug in the Crystal compiler.
Possibly the callbacks are putting return values on the stack when the function signature has a void return type.
Maybe this is misaligning the stack, I'm not sure.